# Implement a more responsive compilation thread in the the language server.