--- a/sway-core/src/ir_generation/const_eval.rs
+++ b/sway-core/src/ir_generation/const_eval.rs
@@ -1140,6 +1140,7 @@ mod tests {
             core_lib,
             None,
             "test",
+            None
         );
 
         let (errors, _warnings) = handler.consume();