--- a/sway-core/src/lib.rs
+++ b/sway-core/src/lib.rs
@@ -607,10 +607,9 @@ pub fn compile_to_ast(
     package_name: &str,
     retrigger_compilation: Option<Arc<AtomicBool>>,
 ) -> Result<Programs, ErrorEmitted> {
+    check_should_abort(handler, retrigger_compilation.clone(), 610)?;
     let query_engine = engines.qe();
     let mut metrics = PerformanceData::default();
-
-    check_should_abort(handler, retrigger_compilation.clone(), 613)?;
 
     if let Some(config) = build_config {
         let path = config.canonical_root_module();
@@ -625,6 +624,7 @@ pub fn compile_to_ast(
             let new_handler = Handler::from_parts(warnings, errors);
             handler.append(new_handler);
 
+            eprintln!("re-using cached prgram data, returning from compilation");
             return Ok(entry.programs);
         };
     }
@@ -963,13 +963,13 @@ fn module_return_path_analysis(
 }
 
 fn check_should_abort(handler: &Handler, retrigger_compilation: Option<Arc<AtomicBool>>, line: u32) -> Result<(), ErrorEmitted> {
-    eprintln!("check_should_abort at line {}.", line);
+    //eprintln!("check_should_abort at line {}.", line);
     if let Some(ref retrigger_compilation) = retrigger_compilation {
         if retrigger_compilation.load(Ordering::SeqCst) {
             eprintln!("Aborting compilation due to retrigger as line {}.", line);
             return Err(handler.cancel());
         } else {
-            eprintln!("Continuing compilation at line {}.", line);
+            //eprintln!("Continuing compilation at line {}.", line);
         }
     } else {
         eprintln!("retrigger_compilation is None at line {}.", line);