--- a/sway-core/src/lib.rs
+++ b/sway-core/src/lib.rs
@@ -35,8 +35,8 @@ use std::collections::hash_map::DefaultHasher;
 use std::collections::HashMap;
 use std::hash::{Hash, Hasher};
 use std::path::{Path, PathBuf};
-use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::Arc;
 use sway_ast::AttributeDecl;
 use sway_error::handler::{ErrorEmitted, Handler};
 use sway_ir::{
@@ -617,7 +617,12 @@ pub fn compile_to_ast(
 
         // Check if we can re-use the data in the cache.
         if is_parse_module_cache_up_to_date(engines, &path, include_tests) {
-            let mut entry = query_engine.get_programs_cache_entry(&path).expect(&format!("unable to find entry in cache at path {:?}", &path));
+            let mut entry = query_engine
+                .get_programs_cache_entry(&path)
+                .expect(&format!(
+                    "unable to find entry in cache at path {:?}",
+                    &path
+                ));
             entry.programs.metrics.reused_modules += 1;
 
             let (warnings, errors) = entry.handler_data;
@@ -962,7 +967,11 @@ fn module_return_path_analysis(
     }
 }
 
-fn check_should_abort(handler: &Handler, retrigger_compilation: Option<Arc<AtomicBool>>, line: u32) -> Result<(), ErrorEmitted> {
+fn check_should_abort(
+    handler: &Handler,
+    retrigger_compilation: Option<Arc<AtomicBool>>,
+    line: u32,
+) -> Result<(), ErrorEmitted> {
     if let Some(ref retrigger_compilation) = retrigger_compilation {
         if retrigger_compilation.load(Ordering::SeqCst) {
             eprintln!("Aborting compilation due to retrigger as line {}.", line);
@@ -975,7 +984,6 @@ fn check_should_abort(handler: &Handler, retrigger_compilation: Option<Arc<Atomi
     }
     Ok(())
 }
-
 
 #[test]
 fn test_basic_prog() {