--- a/sway-core/src/query_engine/mod.rs
+++ b/sway-core/src/query_engine/mod.rs
@@ -45,10 +45,19 @@ pub struct ProgramsCacheEntry {
 
 pub type ProgramsCacheMap = HashMap<ModulePath, ProgramsCacheEntry>;
 
-#[derive(Debug, Default, Clone)]
+#[derive(Debug, Default)]
 pub struct QueryEngine {
-    parse_module_cache: Arc<RwLock<ModuleCacheMap>>,
-    programs_cache: Arc<RwLock<ProgramsCacheMap>>,
+    parse_module_cache: RwLock<ModuleCacheMap>,
+    programs_cache: RwLock<ProgramsCacheMap>,
+}
+
+impl Clone for QueryEngine {
+    fn clone(&self) -> Self {
+        Self {
+            parse_module_cache: RwLock::new(self.parse_module_cache.read().unwrap().clone()),
+            programs_cache: RwLock::new(self.programs_cache.read().unwrap().clone()),
+        }
+    }
 }
 
 impl QueryEngine {