--- a/sway-core/src/query_engine/mod.rs
+++ b/sway-core/src/query_engine/mod.rs
@@ -54,18 +54,8 @@ pub struct QueryEngine {
 impl Clone for QueryEngine {
     fn clone(&self) -> Self {
         Self {
-            parse_module_cache: RwLock::new(
-                self.parse_module_cache
-                    .read()
-                    .expect("Lock is poisoned")
-                    .clone(),
-            ),
-            programs_cache: RwLock::new(
-                self.programs_cache
-                    .read()
-                    .expect("Lock is poisoned")
-                    .clone(),
-            ),
+            parse_module_cache: RwLock::new(self.parse_module_cache.read().unwrap().clone()),
+            programs_cache: RwLock::new(self.programs_cache.read().unwrap().clone()),
         }
     }
 }