… Cargo.lock
… sway-core/src/type_system/engine.rs
… sway-error/src/handler.rs
… sway-lsp/Cargo.toml
… sway-lsp/benches/lsp_benchmarks/compile.rs
… sway-lsp/benches/lsp_benchmarks/mod.rs
… sway-lsp/src/capabilities/code_actions/mod.rs
… sway-lsp/src/capabilities/hover/mod.rs
… sway-lsp/src/capabilities/inlay_hints.rs
… sway-lsp/src/capabilities/rename.rs
… sway-lsp/src/capabilities/semantic_tokens.rs
… sway-lsp/src/core/session.rs
… sway-lsp/src/error.rs
… sway-lsp/src/handlers/notification.rs
… sway-lsp/src/handlers/request.rs
… sway-lsp/src/lib.rs
… sway-lsp/src/server.rs
… sway-lsp/src/server_state.rs
… sway-lsp/tests/fixtures/benchmark/Forc.toml
… sway-lsp/tests/integration/lsp.rs
… sway-lsp/tests/lib.rs
… sway-lsp/tests/utils/Cargo.toml
… sway-types/src/source_engine.rs
… test/src/ir_generation/mod.rs