## Description
This PR implements options 2 & 4 from #4706 in order to drastically improve responsiveness to did change events in the server. We now spawn a dedicated compilation thread to respond and process events in parallel. The compiler now periodically checks if the server has requested it bail out to start compilation on a newer event. I've tried to limit the number of places we perform this check to reduce code "noise". 

We also clone the `Engines` before compiling and write the result back to the `Engines` held by the server on a successful compilation. This is to avoid the `Engines` becoming corrupted if the server requests the compiler to bail out early.  

Some other small changes:
* Diagnostics are now only published on didSave events
* Semantic tokens are only computed for the visible range and not the entire document

Here are the before and after results when  running the `did_change_stress_test` when responding to 400 compilation requests on the benchmark project: 🚀 

Metric | Before | After | Speedup Factor 
-- | -- | -- | -- 
Each Request Time | 78.265ms | 360.03µs | 217x 
Total Time | 31.7856242s | 1.54047762s | 20x 

Previously, we had to wait until compilation had finished before we could respond to incoming events. Now that compilation is done in parallel, we can respond to new events (and set the re-trigger compilation flag) as quickly as possible. This leads to a pretty noticeable improvement to the UX when typing in the editor. 

closes #4706

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand areas.
- [ ] I have updated the documentation where relevant (API docs, the reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my feature works.
- [ ] I have added (or requested a maintainer to add) the necessary `Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs Code Review Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.